wip: decompose folder evidence storage
